% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\docType{class}
\name{Font-class}
\alias{Font}
\alias{Font-class}
\title{S4 class for font objects}
\description{
S4 class for font objects
}
\section{Slots}{

\describe{
\item{\code{fontname}}{length-one character vector}
}}

