% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontFamily.R
\name{font_family}
\alias{font_family}
\alias{createFontFamily}
\title{Create font family}
\usage{
font_family(fontname, reg_wt = "Regular", bold_wt = "Bold")
}
\arguments{
\item{fontname}{Name of font}

\item{reg_wt}{Weight to use for regular}

\item{bold_wt}{Weight to use for bold}
}
\description{
Create font family.
}
