% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_font.R
\name{get_font}
\alias{get_font}
\title{Get font object}
\usage{
get_font(fontname = c("FiraSans", "FiraSansCondensed", "FiraMono",
  "CooperHewitt"), weight = "Regular", shape = "Upright")
}
\arguments{
\item{fontname}{Name of font as string.}

\item{weight}{Weight of font to retrieve.}

\item{shape}{Either \code{\"Upright\"} or \code{\"Italic\"}}
}
\description{
Get font object
}
